@@ -18,6 +18,7 @@ import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry.Registrar
 import java.io.File
+import java.net.URLEncoder
 
 class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
     private lateinit var channel: MethodChannel
@@ -44,8 +45,10 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
             val attributionURL: String? = call.argument("attributionURL")
             val file =  File(activeContext!!.cacheDir,stickerImage)
             val stickerImageFile = FileProvider.getUriForFile(activeContext!!, activeContext!!.applicationContext.packageName + ".com.shekarmudaliyar.social_share", file)
+            val appId: String? = call.argument("appId")
 
             val intent = Intent("com.instagram.share.ADD_TO_STORY")
+            
             intent.type = "image/*"
             intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
@@ -56,10 +59,10 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
                 val backgroundImageFile = FileProvider.getUriForFile(activeContext!!, activeContext!!.applicationContext.packageName + ".com.shekarmudaliyar.social_share", backfile)
                 intent.setDataAndType(backgroundImageFile,"image/*")
             }
+            intent.putExtra("com.instagram.platform.extra.APPLICATION_ID", appId)
             intent.putExtra("content_url", attributionURL)
             intent.putExtra("top_background_color", backgroundTopColor)
             intent.putExtra("bottom_background_color", backgroundBottomColor)
-            Log.d("", activity!!.toString())
             // Instantiate activity and verify it will resolve implicit intent
             activity!!.grantUriPermission("com.instagram.android", stickerImageFile, Intent.FLAG_GRANT_READ_URI_PERMISSION)
             if (activity!!.packageManager.resolveActivity(intent, 0) != null) {
@@ -71,23 +74,33 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
         } else if (call.method == "shareFacebookStory") {
             //share on facebook story
             val stickerImage: String? = call.argument("stickerImage")
+            val backgroundImage: String? = call.argument("backgroundImage")
+
+            val file =  File(activeContext!!.cacheDir,stickerImage)
+            val stickerImageFile = FileProvider.getUriForFile(activeContext!!, activeContext!!.applicationContext.packageName + ".com.shekarmudaliyar.social_share", file)
+
             val backgroundTopColor: String? = call.argument("backgroundTopColor")
             val backgroundBottomColor: String? = call.argument("backgroundBottomColor")
             val attributionURL: String? = call.argument("attributionURL")
             val appId: String? = call.argument("appId")
-
-            val file =  File(activeContext!!.cacheDir,stickerImage)
-            val stickerImageFile = FileProvider.getUriForFile(activeContext!!, activeContext!!.applicationContext.packageName + ".com.shekarmudaliyar.social_share", file)
+            
             val intent = Intent("com.facebook.stories.ADD_TO_STORY")
             intent.type = "image/*"
             intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+
+            if (backgroundImage!=null) {
+                //check if background image is also provided
+                val backfile =  File(activeContext!!.cacheDir,backgroundImage)
+                val backgroundImageFile = FileProvider.getUriForFile(activeContext!!, activeContext!!.applicationContext.packageName + ".com.shekarmudaliyar.social_share", backfile)
+                intent.setDataAndType(backgroundImageFile,"image/*")
+            }
+
             intent.putExtra("com.facebook.platform.extra.APPLICATION_ID", appId)
             intent.putExtra("interactive_asset_uri", stickerImageFile)
             intent.putExtra("content_url", attributionURL)
             intent.putExtra("top_background_color", backgroundTopColor)
             intent.putExtra("bottom_background_color", backgroundBottomColor)
-            Log.d("", activity!!.toString())
             // Instantiate activity and verify it will resolve implicit intent
             activity!!.grantUriPermission("com.facebook.katana", stickerImageFile, Intent.FLAG_GRANT_READ_URI_PERMISSION)
             if (activity!!.packageManager.resolveActivity(intent, 0) != null) {
@@ -121,7 +134,7 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
             chooserIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
 
             activeContext!!.startActivity(chooserIntent)
-            result.success(true)
+            result.success("success")
 
         } else if (call.method == "copyToClipboard") {
             //copies content onto the clipboard
@@ -129,7 +142,7 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
             val clipboard =context!!.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
             val clip = ClipData.newPlainText("", content)
             clipboard.setPrimaryClip(clip)
-            result.success(true)
+            result.success("success")
         } else if (call.method == "shareWhatsapp") {
             //shares content on WhatsApp
             val content: String? = call.argument("content")
@@ -139,9 +152,9 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
             whatsappIntent.putExtra(Intent.EXTRA_TEXT, content)
             try {
                 activity!!.startActivity(whatsappIntent)
-                result.success("true")
+                result.success("success")
             } catch (ex: ActivityNotFoundException) {
-                result.success("false")
+                result.success("error")
             }
         } else if (call.method == "shareSms") {
             //shares content on sms
@@ -153,24 +166,23 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
             intent.putExtra("sms_body", content)
             try {
                 activity!!.startActivity(intent)
-                result.success("true")
+                result.success("success")
             } catch (ex: ActivityNotFoundException) {
-                result.success("false")
+                result.success("error")
             }
         } else if (call.method == "shareTwitter") {
             //shares content on twitter
             val text: String? = call.argument("captionText")
-            val url: String? = call.argument("url")
-            val trailingText: String? = call.argument("trailingText")
-            val urlScheme = "http://www.twitter.com/intent/tweet?text=$text$url$trailingText"
-            Log.d("log",urlScheme)
+            val urlScheme = "http://www.twitter.com/intent/tweet?text=${URLEncoder.encode(text, Charsets.UTF_8.name())}"
+            Log.d("", urlScheme)
+
             val intent = Intent(Intent.ACTION_VIEW)
             intent.data = Uri.parse(urlScheme)
             try {
                 activity!!.startActivity(intent)
-                result.success("true")
+                result.success("success")
             } catch (ex: ActivityNotFoundException) {
-                result.success("false")
+                result.success("error")
             }
         }
         else if (call.method == "shareTelegram") {
@@ -182,9 +194,9 @@ class SocialSharePlugin:FlutterPlugin, MethodCallHandler, ActivityAware {
             telegramIntent.putExtra(Intent.EXTRA_TEXT, content)
             try {
                 activity!!.startActivity(telegramIntent)
-                result.success("true")
+                result.success("success")
             } catch (ex: ActivityNotFoundException) {
-                result.success("false")
+                result.success("error")
             }
         }
         else if (call.method == "checkInstalledApps") {
