@@ -5,10 +5,10 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
+// import 'package:flutter/material.dart';
+// import 'package:flutter_test/flutter_test.dart';
 
-import 'package:social_share_example/main.dart';
+// import 'package:social_share_example/main.dart';
 
 void main() {
   // testWidgets('Verify Platform version', (WidgetTester tester) async {
